+++ src/react/hooks.js
@@ -1,0 +1,82 @@
+const { useContext, useEffect, useState, useRef } = require('react')
+const { AtomContext, differ } = require('../react')
+const raf = require('../raf')
+
+const isServer = typeof navigator === 'undefined'
+
+function useAtom (selector, options = {}) {
+  selector = selector || (state => state)
+  const { sync = false, pure = true, observe = !isServer } = options
+  const schedule = sync ? fn => () => fn() : raf
+  const { atom } = useContext(AtomContext)
+  assert(atom, 'No atom found in context, did you forget to wrap your app in <Provider atom={atom} />?')
+  const [mappedProps, setMappedProps] = useState(selector(atom.get()))
+  const ref = useRef({ mappedProps, atom })
+
+  // keep track of rendering order
+  // this is important for correctness – parent must rerender first
+  // and performance – parent rerendering children should cancel
+  // children's scheduled rerenders
+  if (!ref.current.order) {
+    atom._i = atom._i ? atom._i + 1 : 1
+    ref.current.order = atom._i
+  }
+
+  useEffect(() => {
+    if (!observe) return
+    ref.current.unobserve = atom.observe(() => {
+      // store updates happening in rapid sequence
+      // get cancelled and rescheduled
+      if (ref.current.cancelUpdate) {
+        ref.current.cancelUpdate()
+        ref.current.cancelUpdate = null
+      }
+      // schedule an update
+      ref.current.cancelUpdate = schedule(() => {
+        ref.current.cancelUpdate = null
+        const nextMappedProps = selector(atom.get())
+        if (!pure || differ(ref.current.mappedProps, nextMappedProps)) {
+          setMappedProps(nextMappedProps)
+        }
+      })()
+    }, ref.current.order)
+    return () => {
+      ref.current.unobserve && ref.current.unobserve()
+      if (ref.current.cancelUpdate) {
+        ref.current.cancelUpdate()
+        ref.current.cancelUpdate = null
+      }
+    }
+  }, [atom])
+
+  // cancel a scheduled update on render
+  // we're being rendered right now
+  // so don't need to render again later
+  if (ref.current.cancelUpdate) {
+    ref.current.cancelUpdate()
+    ref.current.cancelUpdate = null
+  }
+
+  // always return fresh mapped props, in case
+  // this is a parent rerendering children
+  ref.current.mappedProps = selector(atom.get())
+  return ref.current.mappedProps
+}
+
+function useActions () {
+  const { atom } = useContext(AtomContext)
+  return atom.actions
+}
+
+function useDispatch () {
+  const { atom } = useContext(AtomContext)
+  return atom.dispatch
+}
+
+function assert (cond, error) {
+  if (!cond) {
+    throw new Error(error)
+  }
+}
+
+module.exports = { useAtom, useActions, useDispatch }